+++ src/Faker/Factory.php
@@ -20,7 +20,8 @@ class Factory implements ContainerAwareInterface
             new DrupalEntity(
                 $generator,
                 $this->container,
-                $this->container->get('wmmodel.factory.model')
+                $this->container->get('entity_type.manager'),
+                $this->container->get('entity_type.repository')
             )
         );