+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2022-01-11
+### Changed
+- Increase minimum Drupal core version to 9.3 due to entity bundle class support
+- Increase minimum PHP requirement to 7.3 due to Drupal core PHP requirement
+
+### Removed
+- Remove wieni/wmmodel dependency
 
 ## [1.1.1] - 2020-12-22
 ### Added